+++ src/main/java/org/team4/hanzip/global/security/jwt/JwtValidator.java
@@ -1,0 +1,37 @@
+package org.team4.hanzip.global.security.jwt;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import jakarta.annotation.PostConstruct;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.util.Date;
+
+@Component
+public class JwtValidator {
+    @Value("${jwt.secret}")
+    private String secret;
+    private SecretKey key;
+
+    @PostConstruct
+    void Init(){
+        key = Keys.hmacShaKeyFor(secret.getBytes());
+    }
+
+    public Boolean validateToken(String token) {
+        try {
+            Jws<Claims> claims = Jwts.parser()
+                    .setSigningKey(key)
+                    .build()
+                    .parseClaimsJws(token);
+
+            return claims.getBody().getExpiration().after(new Date());
+        } catch (Exception e) {
+            return false;
+        }
+    }
+}
